refactor: 💡 Parser - Lexical Analysis
Refactoring into different files
