+++ testing/run_test.sh
@@ -4,6 +4,9 @@ export LD_LIBRARY_PATH=./criterion-2.4.2/lib:$LD_LIBRARY_PATH
 gcc -I \
 	./criterion-2.4.2/include \
 	minishell_test.c \
+	minishell_test_utils.c \
+	minishell_test_builtin.c \
+	minishell_test_quote.c \
 	../src/parser_token_build.c \
 	../src/parser_token_utils.c \
 	../src/parser_token_define.c \
